Completed Tasks (#18)
* login with email, non-unique username, correct timestamps for daily words and user table

* updated changelog table, user table has update  by/at and created by, application file restored to og

* ajay's changes updated

* applicationand gitignore file reverted to og

* removed commented code and updated time for daily response working fine

+++ src/main/java/com/mas/dashboard/dto/TaskRatingDto.java
@@ -12,7 +12,7 @@ public class TaskRatingDto {
 
   private Long studentId;
 
-  private Long rating;
+  private float rating;
 
   private Boolean deleted;
 }
